Fix:classify signup function to UserController, getAllUsers function to UserDataController
